+++ feature/gameschedule/data/src/main/java/com/pavan/gamebay/feature/gameschedule/data/local/mapper/GameScheduleRoomMapper.kt
@@ -1,6 +1,5 @@
 package com.pavan.gamebay.feature.gameschedule.data.local.mapper
 
-import android.R.attr.label
 import com.pavan.gamebay.core.domain.DefaultDispatcher
 import com.pavan.gamebay.core.domain.IMapper
 import com.pavan.gamebay.feature.gameschedule.data.local.entities.GameEntity
@@ -9,22 +8,41 @@ import com.pavan.gamebay.feature.gameschedule.data.local.entities.ScheduleWithDe
 import com.pavan.gamebay.feature.gameschedule.data.local.entities.TeamEntity
 import com.pavan.gamebay.feature.gameschedule.domain.models.Game
 import com.pavan.gamebay.feature.gameschedule.domain.models.GameDate
+import com.pavan.gamebay.feature.gameschedule.domain.models.GameOutcome
 import com.pavan.gamebay.feature.gameschedule.domain.models.GameSection
+import com.pavan.gamebay.feature.gameschedule.domain.models.GameType
 import com.pavan.gamebay.feature.gameschedule.domain.models.Schedule
 import com.pavan.gamebay.feature.gameschedule.domain.models.Team
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
+/**
+ * Mapper class to convert between `ScheduleWithDetails` and `Schedule` domain models.
+ *
+ * @property defaultDispatcher The default coroutine dispatcher for executing mapping operations.
+ */
 class GameScheduleRoomMapper @Inject constructor(
     @DefaultDispatcher private val defaultDispatcher: CoroutineDispatcher
 ) : IMapper<ScheduleWithDetails, Schedule> {
+
+    /**
+     * Maps a `ScheduleWithDetails` object to a `Schedule` object.
+     *
+     * @param input The `ScheduleWithDetails` object to be mapped.
+     * @return The mapped `Schedule` object.
+     */
     override suspend fun map(input: ScheduleWithDetails): Schedule {
         return withContext(defaultDispatcher) {
             input.toDomain()
         }
     }
 
+    /**
+     * Extension function to map `ScheduleWithDetails` to `Schedule`.
+     *
+     * @return The mapped `Schedule` object.
+     */
     private fun ScheduleWithDetails.toDomain() = Schedule(
         team = this.schedule?.team.toDomain(),
         defaultGameId = this.schedule?.defaultGameId?.toLong() ?: 0L,
@@ -33,6 +51,11 @@ class GameScheduleRoomMapper @Inject constructor(
         }
     )
 
+    /**
+     * Extension function to map `GameSectionWithGames` to `GameSection`.
+     *
+     * @return The mapped `GameSection` object.
+     */
     private fun GameSectionWithGames.toDomain() = GameSection(
         heading = this.gameSection?.heading ?: "",
         games = this.games.map {
@@ -40,6 +63,11 @@ class GameScheduleRoomMapper @Inject constructor(
         }
     )
 
+    /**
+     * Extension function to map `GameEntity` to `Game`.
+     *
+     * @return The mapped `Game` object.
+     */
     private fun GameEntity.toDomain() = Game(
         id = id,
         week = week,
@@ -55,7 +83,11 @@ class GameScheduleRoomMapper @Inject constructor(
         opponentTeam = opponent.toDomain()
     )
 
-
+    /**
+     * Extension function to map `TeamEntity` to `Team`.
+     *
+     * @return The mapped `Team` object.
+     */
     private fun TeamEntity?.toDomain() = Team(
         triCode = this?.triCode ?: "",
         fullName = this?.fullName ?: "",
@@ -68,6 +100,11 @@ class GameScheduleRoomMapper @Inject constructor(
         primaryColor = this?.primaryColor ?: ""
     )
 
+    /**
+     * Extension function to map `GameEntity` to `GameDate`.
+     *
+     * @return The mapped `GameDate` object.
+     */
     private fun GameEntity?.toGameDate(): GameDate {
         return this?.let {
             GameDate(
@@ -80,17 +117,27 @@ class GameScheduleRoomMapper @Inject constructor(
         )
     }
 
+    /**
+     * Extension function to map a nullable `String` to `GameOutcome`.
+     *
+     * @return The mapped `GameOutcome` object.
+     */
     private fun String?.asGameOutcome() = when (this) {
-        "W" -> com.pavan.gamebay.feature.gameschedule.domain.models.GameOutcome.WIN
-        "L" -> com.pavan.gamebay.feature.gameschedule.domain.models.GameOutcome.LOSS
-        "T" -> com.pavan.gamebay.feature.gameschedule.domain.models.GameOutcome.TIE
-        else -> com.pavan.gamebay.feature.gameschedule.domain.models.GameOutcome.NONE
+        "W" -> GameOutcome.WIN
+        "L" -> GameOutcome.LOSS
+        "T" -> GameOutcome.TIE
+        else -> GameOutcome.NONE
     }
 
+    /**
+     * Extension function to map a nullable `String` to `GameType`.
+     *
+     * @return The mapped `GameType` object.
+     */
     private fun String?.asGameType() = when (this) {
-        "F" -> com.pavan.gamebay.feature.gameschedule.domain.models.GameType.FINAL
-        "B" -> com.pavan.gamebay.feature.gameschedule.domain.models.GameType.BYE
-        "S" -> com.pavan.gamebay.feature.gameschedule.domain.models.GameType.SCHEDULED
-        else -> com.pavan.gamebay.feature.gameschedule.domain.models.GameType.NONE
+        "F" -> GameType.FINAL
+        "B" -> GameType.BYE
+        "S" -> GameType.SCHEDULED
+        else -> GameType.NONE
     }
 }
