+++ feature/gameschedule/domain/src/main/java/com/pavan/gamebay/feature/gameschedule/domain/models/GameDate.kt
@@ -1,0 +1,12 @@
+package com.pavan.gamebay.feature.gameschedule.domain.models
+
+/**
+ * Data class representing a game date.
+ *
+ * @property dateText A string representation of the date.
+ * @property dateTime A string representation of the date and time.
+ */
+data class GameDate(
+    val dateText: String,
+    val dateTime: String,
+)
