+++ feature/gameschedule/domain/src/main/java/com/pavan/gamebay/feature/gameschedule/domain/models/GameOutcome.kt
@@ -1,0 +1,26 @@
+package com.pavan.gamebay.feature.gameschedule.domain.models
+
+/**
+ * Enum class representing the possible outcomes of a game.
+ */
+enum class GameOutcome {
+    /**
+     * Represents a win outcome.
+     */
+    WIN,
+
+    /**
+     * Represents a loss outcome.
+     */
+    LOSS,
+
+    /**
+     * Represents a tie outcome.
+     */
+    TIE,
+
+    /**
+     * Represents no outcome.
+     */
+    NONE
+}
