+++ feature/gameschedule/domain/src/main/java/com/pavan/gamebay/feature/gameschedule/domain/models/GameType.kt
@@ -1,0 +1,26 @@
+package com.pavan.gamebay.feature.gameschedule.domain.models
+
+/**
+ * Enum class representing different types of game statuses.
+ */
+enum class GameType {
+    /**
+     * Represents a final game.
+     */
+    FINAL,
+
+    /**
+     * Represents a bye game.
+     */
+    BYE,
+
+    /**
+     * Represents a scheduled game.
+     */
+    SCHEDULED,
+
+    /**
+     * Represents no game.
+     */
+    NONE
+}
