+++ gradle/libs.versions.toml
@@ -62,6 +62,9 @@ androidx-room-compiler = { group = "androidx.room", name = "room-compiler", vers
 androidx-appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
 material3 = { group = "androidx.compose.material3", name = "material3" }
 
+# Testing dependencies
+mockk = { group = "io.mockk", name = "mockk", version = "1.12.0" }
+
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
 android-library = { id = "com.android.library", version.ref = "agp" }
